Fix getting multiple entries in syslog.conf
git-svn-id: https://svn.code.sf.net/p/xcat/code/xcat-core/trunk@581 8638fb3e-16cb-4fca-ae20-7b5d299a9bcd

+++ xCATsn/xCATsn.spec
@@ -59,15 +59,17 @@ if [ "$1" = "1" ]; then #Only if installing for the first time..
 # makes it a service node
   touch /etc/xCATSN    
 # setup syslog
-  cp /etc/syslog.conf /etc/syslog.conf.ORIG
-  echo "*.debug   /var/log/localmessages" > /etc/test.tmp 
-  echo "*.crit   /var/log/localmessages" >> /etc/test.tmp 
+  if [ ! -r /etc/syslog.conf.XCATORIG ]; then
+  cp /etc/syslog.conf /etc/syslog.conf.XCATORIG
+  echo "*.debug   /var/log/messages" > /etc/test.tmp 
+  echo "*.crit   /var/log/messages" >> /etc/test.tmp 
   cat /etc/test.tmp >> /etc/syslog.conf
   rm /etc/test.tmp
-  touch /var/log/localmessages
+  touch /var/log/messages
 
   /etc/rc.d/init.d/syslog stop
   /etc/rc.d/init.d/syslog start
+  fi
